+++ src/components/PupUp/AdminPopUp.js
@@ -5,6 +5,7 @@ import { observer, inject } from "mobx-react";
 import { makeStyles } from "@material-ui/core/styles";
 import TextField from "@material-ui/core/TextField";
 import dayjs from "dayjs";
+import Swal from 'sweetalert2'
 import "./AdminPopUp.css";
 const useStyles = makeStyles((theme) => ({
   container: {
@@ -64,11 +65,27 @@ const AdminPopUp = inject("ScheduleStore")(
       debugger;
       handleClose();
     };
-    const handleDelete = () => {
-      props.ScheduleStore.deleteSchedule(schedule);
-      props.handleCloseModal();
-    };
 
+    const handleDelete=()=>{
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+    }).then(async (result) => {
+        if (result.isConfirmed) {
+          props.ScheduleStore.deleteSchedule(schedule);
+            Swal.fire(
+                'Deleted!',
+                `The data has been deleted. 😔`,
+                'success'
+            )}
+            props.handleCloseModal()
+          })
+    }
     const handleClose = () => {
       props.handleCloseModal();
     };
